Clean args files stuff

+++ include/torchrl/algorithms/ppo/PPO.hpp
@@ -12,7 +12,7 @@ public:
     PPO(VectorizedEnv& env_, const PPOArgs& args);
     ~PPO();
 
-    void Learn(const size_t total_timesteps);
+    void Learn(const uint64_t total_timesteps);
 
 private:
     /// @brief Use the policy to play in the env and store the data in buffer
